Feature/product quantity

+++ src/main/java/rastle/dev/rastle_backend/domain/payment/dto/PaymentDTO.java
@@ -53,6 +53,6 @@ public class PaymentDTO {
     @Builder
     public static class PaymentPrepareResponse {
         String merchant_uid;
-        Integer number;
+        Long number;
     }
 }
